+++ src/timessquare/domain/pageparameters/__init__.py
@@ -6,9 +6,10 @@ from ._booleanparameter import BooleanParameterSchema
 from ._datedynamicdefault import DYNAMIC_DATE_PATTERN, DateDynamicDefault
 from ._dateparameter import DateParameterSchema
 from ._datetimeparameter import DatetimeParameterSchema
+from ._dayobsdateparameter import DayObsDateParameterSchema
+from ._dayobsparameter import DayObsParameterSchema
 from ._integerparameter import IntegerParameterSchema
 from ._numberparameter import NumberParameterSchema
-from ._obsdateparameter import ObsDateParameterSchema
 from ._pageparameters import (
     PageParameters,
     create_and_validate_parameter_schema,
@@ -23,9 +24,10 @@ __all__ = [
     "DateDynamicDefault",
     "DateParameterSchema",
     "DatetimeParameterSchema",
+    "DayObsDateParameterSchema",
+    "DayObsParameterSchema",
     "IntegerParameterSchema",
     "NumberParameterSchema",
-    "ObsDateParameterSchema",
     "PageParameterSchema",
     "PageParameters",
     "StringParameterSchema",